refactor: Praise, Mate Domain 분리
