+++ src/main/java/com/gloddy/server/estimate/application/EstimateService.java
@@ -3,8 +3,8 @@ package com.gloddy.server.estimate.application;
 import com.gloddy.server.core.event.reliability.ReliabilityEventPublisher;
 import com.gloddy.server.core.event.reliability.ReliabilityScoreUpdateEvent;
 import com.gloddy.server.estimate.domain.dto.EstimateRequest;
-import com.gloddy.server.estimate.application.mate.MateSaveService;
-import com.gloddy.server.estimate.application.praise.PraiseService;
+import com.gloddy.server.mate.application.MateSaveService;
+import com.gloddy.server.praise.application.PraiseService;
 import com.gloddy.server.user_group.application.UserGroupFindService;
 import com.gloddy.server.user_group.application.UserGroupUpdateService;
 import com.gloddy.server.reliability.domain.vo.ScoreType;
