+++ src/main/java/com/gloddy/server/mate/application/MateGetService.java
@@ -1,8 +1,8 @@
-package com.gloddy.server.estimate.application.mate;
+package com.gloddy.server.mate.application;
 
 import com.gloddy.server.auth.domain.User;
-import com.gloddy.server.estimate.domain.Mate;
-import com.gloddy.server.estimate.infra.repository.MateJpaRepository;
+import com.gloddy.server.mate.domain.Mate;
+import com.gloddy.server.mate.infra.repository.MateJpaRepository;
 import com.gloddy.server.user.domain.handler.UserQueryHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
