+++ src/main/java/com/gloddy/server/praise/domain/service/ActivePraiseStrategy.java
@@ -1,4 +1,4 @@
-package com.gloddy.server.estimate.domain.vo.strategy;
+package com.gloddy.server.praise.domain.service;
 
 import com.gloddy.server.user_group.domain.UserGroup;
