+++ src/main/java/com/gloddy/server/user/application/UserService.java
@@ -5,9 +5,9 @@ import com.gloddy.server.auth.domain.vo.kind.Personality;
 import com.gloddy.server.core.error.handler.errorCode.ErrorCode;
 import com.gloddy.server.core.error.handler.exception.PraiseBusinessException;
 import com.gloddy.server.core.utils.TimeUtil;
-import com.gloddy.server.estimate.domain.Praise;
-import com.gloddy.server.estimate.infra.repository.MateJpaRepository;
-import com.gloddy.server.estimate.infra.repository.PraiseJpaRepository;
+import com.gloddy.server.praise.domain.Praise;
+import com.gloddy.server.mate.infra.repository.MateJpaRepository;
+import com.gloddy.server.praise.infra.repository.PraiseJpaRepository;
 import com.gloddy.server.user.domain.handler.UserQueryHandler;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
