+++ src/test/java/com/gloddy/server/acceptance/reliability/UpdateReliabilityByEstimateTest.java
@@ -4,9 +4,9 @@ import com.gloddy.server.auth.domain.User;
 import com.gloddy.server.common.reliability.ReliabilityApiTest;
 import com.gloddy.server.core.event.reliability.ReliabilityScoreUpdateEvent;
 import com.gloddy.server.estimate.domain.dto.EstimateRequest;
-import com.gloddy.server.estimate.domain.vo.PraiseValue;
-import com.gloddy.server.estimate.application.mate.MateSaveService;
-import com.gloddy.server.estimate.application.praise.PraiseService;
+import com.gloddy.server.praise.domain.vo.PraiseValue;
+import com.gloddy.server.mate.application.MateSaveService;
+import com.gloddy.server.praise.application.PraiseService;
 import com.gloddy.server.group.domain.Group;
 import com.gloddy.server.user_group.domain.UserGroup;
 import com.gloddy.server.user_group.application.UserGroupUpdateService;
