+++ src/test/java/com/gloddy/server/acceptance/reliability/UpdateReliabilityByPraisedTest.java
@@ -4,8 +4,8 @@ import com.gloddy.server.auth.domain.User;
 import com.gloddy.server.common.reliability.ReliabilityApiTest;
 import com.gloddy.server.core.event.reliability.ReliabilityScoreUpdateEvent;
 import com.gloddy.server.estimate.domain.dto.EstimateRequest;
-import com.gloddy.server.estimate.domain.vo.PraiseValue;
-import com.gloddy.server.estimate.application.mate.MateSaveService;
+import com.gloddy.server.praise.domain.vo.PraiseValue;
+import com.gloddy.server.mate.application.MateSaveService;
 import com.gloddy.server.group.domain.Group;
 import com.gloddy.server.user_group.application.UserGroupUpdateService;
 import com.gloddy.server.reliability.domain.Reliability;
