+++ src/test/java/com/gloddy/server/common/BaseApiTest.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.gloddy.server.auth.domain.User;
 import com.gloddy.server.auth.domain.vo.kind.Personality;
 import com.gloddy.server.auth.jwt.JwtTokenBuilder;
-import com.gloddy.server.estimate.domain.Praise;
-import com.gloddy.server.estimate.infra.repository.PraiseJpaRepository;
+import com.gloddy.server.praise.domain.Praise;
+import com.gloddy.server.praise.infra.repository.PraiseJpaRepository;
 import com.gloddy.server.reliability.domain.Reliability;
 import com.gloddy.server.reliability.domain.handler.ReliabilityQueryHandler;
 import com.gloddy.server.reliability.infra.repository.ReliabilityRepository;
