+++ src/test/java/com/gloddy/server/common/reliability/ReliabilityApiTest.java
@@ -5,8 +5,8 @@ import com.gloddy.server.common.BaseApiTest;
 import com.gloddy.server.estimate.domain.dto.EstimateRequest;
 import com.gloddy.server.estimate.domain.dto.MateDto;
 import com.gloddy.server.estimate.domain.dto.PraiseDto;
-import com.gloddy.server.estimate.domain.vo.PraiseValue;
-import com.gloddy.server.estimate.infra.repository.MateJpaRepository;
+import com.gloddy.server.praise.domain.vo.PraiseValue;
+import com.gloddy.server.mate.infra.repository.MateJpaRepository;
 import com.gloddy.server.group.domain.dto.GroupRequest;
 import com.gloddy.server.group.domain.Group;
 import com.gloddy.server.group.domain.vo.UserGroupVO;
