+++ src/test/java/com/gloddy/server/common/userGroup/UserGroupDomainTest.java
@@ -2,7 +2,7 @@ package com.gloddy.server.common.userGroup;
 
 import com.gloddy.server.UnitTest;
 import com.gloddy.server.auth.domain.User;
-import com.gloddy.server.estimate.domain.Praise;
+import com.gloddy.server.praise.domain.Praise;
 import com.gloddy.server.group.domain.Group;
 import com.gloddy.server.user_group.domain.UserGroup;
