+++ src/test/java/com/gloddy/server/unit/userGroup/ReceivePraiseTest.java
@@ -2,8 +2,8 @@ package com.gloddy.server.unit.userGroup;
 
 import com.gloddy.server.auth.domain.User;
 import com.gloddy.server.common.userGroup.UserGroupDomainTest;
-import com.gloddy.server.estimate.domain.Praise;
-import com.gloddy.server.estimate.domain.vo.PraiseValue;
+import com.gloddy.server.praise.domain.Praise;
+import com.gloddy.server.praise.domain.vo.PraiseValue;
 import com.gloddy.server.group.domain.Group;
 import com.gloddy.server.user_group.domain.UserGroup;
 import org.junit.jupiter.api.DisplayName;
